+++ attributes/default.rb
@@ -27,6 +27,7 @@ default['shibboleth-idp']['status_allowed_ips'] = %w{ 127.0.0.1/32 ::1/128 }
 default['shibboleth-idp']['tomcat6_dta_ssl_download_url'] = "https://build.shibboleth.net/nexus/content/repositories/releases/edu/internet2/middleware/security/tomcat6/tomcat6-dta-ssl/1.0.0/tomcat6-dta-ssl-1.0.0.jar"
 default['shibboleth-idp']['template_cookbook'] = "shibboleth-idp"
 default['shibboleth-idp']['ip_tables']['in_interface'] = nil
+default['shibboleth-idp']['custom_webapp_templates'] = nil
 
 case node["platform_family"]
 when "rhel"